==> scripts/0_Surface_area.Rmd <==
---
title: "Surface Area Calculations from Wax Dipping"
author: "HM Putnam, AS Huffmyer"
edited by: "HM Putnam"
date: "20230327"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

## install packages if you dont already have them
if (!require("tidyverse")) install.packages("tidyverse")

# load packages
library(tidyverse)
```

TPC Curve Data
#use wooden dowel standards to calculate surface area curve 
```{r}

#load wax data
wax.data <- read.csv("data/May_TPCBASS_surface_area_data.csv", header=TRUE)

wax.data$delta.mass.g <- wax.data$weight2.g-wax.data$weight1.g
stnds <- subset(wax.data, Sample=="Standard")
plot(stnds$delta.mass.g)

#calculate the surface area of the spherical standards from the diameter
stnds$rad <- stnds$Diameter/2
stnds$surface.area.cm2 <- 4*pi*(stnds$rad)^2

# calculate the curve coefficients for slope and intercept to apply as the standard
stnd.curve <- lm(surface.area.cm2~delta.mass.g, data=stnds)
plot(surface.area.cm2~delta.mass.g, data=stnds)
stnd.curve$coefficients
summary(stnd.curve)$r.squared #r-squared 


```

# May Samples
```{r}
#Calculate surface area using the standard curve
smpls <- subset(wax.data, Sample=="Coral")
smpls$surface.area.cm2 <- stnd.curve$coefficients[2] * smpls$delta.mass.g + stnd.curve$coefficients[1]

#select the samples only
smpls <- smpls %>%
  select(-Sample, -Diameter)

smpls %>%
  count(fragment_ID) %>% arrange(n)

#check the range to make sure your samples fall within the range of the standards
range(smpls$surface.area.cm2)
hist(smpls$surface.area.cm2)
range(stnds$surface.area.cm2)

#Save the output for use in normalization for phys assays
smpls%>%
  select(fragment_ID, surface.area.cm2)%>%
    write_csv("data/TPC/May_TPCBASS_surface.area.calc.csv")
```

# December Samples
```{r}
#Calculate surface area using the standard curve
wax.data <- read.csv("data/Dec_TPCBASS_surface_area_data.csv", header=TRUE)
wax.data$delta.mass.g <- wax.data$weight2.g-wax.data$weight1.g
smpls <- subset(wax.data, Sample=="Coral")
smpls$surface.area.cm2 <- stnd.curve$coefficients[2] * smpls$delta.mass.g + stnd.curve$coefficients[1]

#select the samples only
smpls <- smpls %>%
  select(-Sample, -Diameter)

smpls %>%
  count(fragment_ID) %>% arrange(n)

#check the range to make sure your samples fall within the range of the standards
range(smpls$surface.area.cm2)
hist(smpls$surface.area.cm2)
range(stnds$surface.area.cm2)

#Save the output for use in normalization for phys assays
smpls%>%
  select(fragment_ID, surface.area.cm2)%>%
    write_csv("data/TPC/Dec_TPCBASS_surface.area.calc.csv")
```
