==> scripts/2_Dec_Photo_rate_TPC_fitting.Rmd <==
---
title: "Photosynthesis TPC fitting"
authors: Hollie Putnam 
edited by: Hollie Putnam 
date: 20231008
output: html_document
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

library('rTPC')
library('nls.multstart')
library('tidyverse')
library('broom')
library('gridExtra')
```

Full Data Set
```{r, warning = FALSE}
Photodata <- read.csv("data/Dec_Photo_Rawextracted_rates.csv") #Import data
Photodata$group <-paste0(Photodata$species,"_", Photodata$Temp.Cat)

#plot all raw data
with <- ggplot(Photodata) +
  geom_point(aes(Temp.Cat, micromol.cm2.h, shape = species)) +
  #facet_wrap(~fragment_ID, ncol = 6) +
  theme_bw() +
  theme(legend.position = 'none') +
  scale_color_brewer(type = 'qual', palette = 2) +
  labs(x = 'Temperature (ºC)',
       y = 'Metabolic rate',
       title = 'Dec All')
with
```


# Remove outliers
```{r, warning=FALSE}
#set quantile values
q <- c(0.25, 0.75)

# calculate quantile values by Temperature and Treatment groups
Quants <- Photodata %>%
  group_by(species, Temp.Cat) %>%
  summarize(quant25 = quantile(micromol.cm2.h, probs = q[1]),
            quant75 = quantile(micromol.cm2.h, probs = q[2]),
            IQRbyGroup=IQR(micromol.cm2.h))

# add a group name by pasting Temperature and Treatment
Quants$group <-paste0(Quants$species,"_", Quants$Temp.Cat)

#Calculate Quantile upper and lower ranges
Quants$upper <-  Quants$quant75+1.5*Quants$IQRbyGroup # Upper Range
Quants$lower <- Quants$quant25-1.5*Quants$IQRbyGroup # Lower Range

#join outlier cutoffs with rate data
Photodata <- left_join(Photodata, Quants)

#remove outliers from rates
Photodata <- Photodata %>%
  filter(micromol.cm2.h < upper) %>%
  filter(micromol.cm2.h > lower) 

without <-  ggplot(Photodata) +
  geom_point(aes(Temp.Cat, micromol.cm2.h)) +
  #facet_wrap(~fragment_ID, ncol = 6) +
  theme_bw() +
  theme(legend.position = 'none') +
  scale_color_brewer(type = 'qual', palette = 2) +
  labs(x = 'Temperature (ºC)',
       y = 'Metabolic rate',
       title = 'Dec Outliers removed')

grid.arrange(with, without, ncol=2)


```

TPC fitting 
Padifeld et al **rTPC and nls.multstart: A new pipeline to fit thermal performance curves in r**  
https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13585  

Rezende 2019 model

```{r}
#Photosynthesis
Photodata$temp <- as.numeric(Photodata$Temp.Cat)
Photodata$transformed.rate <- log10(Photodata$micromol.cm2.h+1)

# choose model: Rezende 2019

# get start vals
start_vals <- get_start_vals(Photodata$temp, Photodata$transformed.rate, 
                             model_name = 'rezende_2019')
start_vals[2] <- 0.01

# get limits
low_lims <- get_lower_lims(Photodata$temp, Photodata$transformed.rate, 
                           model_name = 'rezende_2019')
upper_lims <- get_upper_lims(Photodata$temp, Photodata$transformed.rate, 
                             model_name = 'rezende_2019')

#view values
start_vals
low_lims
upper_lims
```

# Individal fitting
```{r}
# fit  chosen model formulation in rTPC
d_fits <- Photodata %>% group_by(fragment_ID) %>% nest() %>%
  mutate(rezende = map(data, ~nls_multstart(transformed.rate~rezende_2019(temp = Temp.Cat, q10, a, b, c),
                        data = .x,
                        iter = c(5,5,5,5),
                        start_lower = start_vals - 1,
                        start_upper = start_vals + 1,
                        lower = low_lims,
                        upper = upper_lims,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


```

```{r}
# create new list column of for high resolution data
d_preds <- mutate(d_fits, new_data = map(data, ~tibble(temp = seq(min(.x$temp), max(.x$temp), length.out = 500)))) %>%
  # get rid of original data column
  select(., -data) %>%
  # stack models into a single column, with an id column for model_name
  pivot_longer(., names_to = 'model_name', values_to = 'fit', rezende) %>%
  # create new list column containing the predictions
  # this uses both fit and new_data list columns
  mutate(preds = map2(fit, new_data, ~augment(.x, new_data = .y))) %>%
  # select only the columns we want to keep
  select(fragment_ID, preds) %>%
  # unlist the preds list column
  unnest(preds)

glimpse(d_preds)



```


```{r}

ggplot(d_preds) +
  geom_line(aes(Temp.Cat, .fitted)) +
  geom_point(aes(Temp.Cat, transformed.rate), mutate(Photodata, fragment_ID = fragment_ID)) +
  facet_wrap(~fragment_ID, ncol = 6) +
  theme_bw() +
  theme(legend.position = 'none') +
  scale_color_brewer(type = 'qual', palette = 2) +
  labs(x = 'Temperature (ºC)',
       y = 'Metabolic rate',
       title = 'Dec fitted thermal performance curves')
```

```{r}
d_params1 <- pivot_longer(d_fits, names_to = 'model_name', values_to = 'fit', rezende) %>%
  mutate(params = map(fit, calc_params),
         rss = map_dbl(fit, deviance)) 
d_params1 %>% arrange(rss)

#### Filter Topt data based on quality of fit?
d_params <- d_params1 %>%
  #filter(rss < 0.01) %>%
  select(fragment_ID, rss, params) %>%
  unnest(params)

glimpse(d_params)

metadata <- read.csv("data/Dec_metadata_POC_TPCBASS.csv")

d_params <- left_join(d_params, metadata, by="fragment_ID")

d_params
write.csv(d_params, file = "data/processed/TPC_Dec_Photo_params_rezende_ind.csv")

d_params_long <- d_params %>%
  pivot_longer(cols = rss:skewness, names_to = "metric", values_to = "value")

d_params_long %>%
ggplot(aes(x=species, y=value, color=species))+
  geom_point()+
  facet_wrap("metric", scales="free_y")+
  theme_bw()

```


# Fit TPCs to species groups instead of individual corals
```{r}
# Function to get fitted values and plot
plot_fit <- function(fits) {
  pgra <- augment(fits$model[[1]])
  pver <- augment(fits$model[[2]])
  d_preds_groups <- bind_rows(`P. grandis` = pgra, `P. verrucosa` = pver, .id = "species")
  plot <- ggplot(d_preds_groups) +
    geom_line(aes(Temp.Cat, .fitted)) +
    geom_point(aes(Temp.Cat, transformed.rate), Photodata) +
    facet_wrap(~species) +
    theme_bw() +
    theme(legend.position = 'none') +
    scale_color_brewer(type = 'qual', palette = 2) +
    labs(x = 'Temperature (ºC)',
         y = 'Metabolic rate',
         title = 'Dec fitted thermal performance curves')
  return(plot)
}

# FUnction to get parameters from fit
get_pars <- function(fits) {
  params <- pivot_longer(fits, names_to = 'model_name', values_to = 'fit', model) %>%
    mutate(params = map(fit, calc_params),
           AIC = map(fit, AIC)) %>%
    select(species, params, AIC) %>%
    unnest(params, AIC)
  params
}

```

```{r fit_many_models, eval = FALSE}
# fit  SPAIN formulation in rTPC
spain_fits <- Photodata %>% group_by(species) %>% nest() %>%
  mutate(model = map(data, ~nls_multstart(transformed.rate~spain_1982(temp = Temp.Cat, a, b, c, r0),
                        data = .x,
                        iter = c(5,5,5,5),
                        start_lower = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'spain_1982') - 1,
                        start_upper = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'spain_1982') + 1,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


plot_fit(spain_fits) + labs(title = "Spain")
get_pars(spain_fits)


# Boatman can't go negative...

# fit Hinshelwood formulation in rTPC
hinshelwood_fits <- Photodata %>% group_by(species) %>% nest() %>%
  mutate(model = map(data, ~nls_multstart(transformed.rate~hinshelwood_1947(temp = Temp.Cat, a,e,b,eh),
                        data = .x,
                        iter = c(5,5,5,5),
                        start_lower = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'hinshelwood_1947') - 1,
                        start_upper = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'hinshelwood_1947') + 1,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


plot_fit(hinshelwood_fits) + labs(title = "Hinshelwood")
get_pars(hinshelwood_fits)



# fit Joehnk formulation in rTPC
joehnk_fits <- Photodata %>% group_by(species) %>% nest() %>%
  mutate(model = map(data, ~nls_multstart(transformed.rate~joehnk_2008(temp = Temp.Cat, rmax, topt, a, b, c),
                        data = .x,
                        iter = c(5,5,5,5,5),
                        start_lower = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'joehnk_2008') - 1,
                        start_upper = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'joehnk_2008') + 1,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


plot_fit(joehnk_fits) + labs(title = "joehnk")
get_pars(joehnk_fits)


# fit Kamykowski formulation in rTPC
kamykowski_fits <- Photodata %>% group_by(species) %>% nest() %>%
  mutate(model = map(data, ~nls_multstart(transformed.rate~kamykowski_1985(temp = Temp.Cat, tmin, tmax, a, b, c),
                        data = .x,
                        iter = c(5,5,5,5,5),
                        start_lower = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'kamykowski_1985') - 1,
                        start_upper = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'kamykowski_1985') + 1,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


plot_fit(kamykowski_fits) + labs(title = "kamykowski")
get_pars(kamykowski_fits)


# fit lactin2 formulation in rTPC
lactin2_fits <- Photodata %>% group_by(species) %>% nest() %>%
  mutate(model = map(data, ~nls_multstart(transformed.rate~lactin2_1995(temp = Temp.Cat, a, b, tmax, delta_t),
                        data = .x,
                        iter = c(5,5,5,5),
                        start_lower = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'lactin2_1995') - 1,
                        start_upper = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'lactin2_1995') + 1,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


plot_fit(lactin2_fits) + labs(title = "lactin2")
get_pars(lactin2_fits)

# fit lrf formulation in rTPC
lrf_fits <- Photodata %>% group_by(species) %>% nest() %>%
  mutate(model = map(data, ~nls_multstart(transformed.rate~lrf_1991(temp = Temp.Cat, rmax, topt, tmin, tmax),
                        data = .x,
                        iter = c(5,5,5,5),
                        start_lower = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'lrf_1991') - 1,
                        start_upper = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'lrf_1991') + 1,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


plot_fit(lrf_fits) + labs(title = "lrf")
get_pars(lrf_fits)

# fit rezende formulation in rTPC
rezende_fits <- Photodata %>% group_by(species) %>% nest() %>%
  mutate(model = map(data, ~nls_multstart(transformed.rate~rezende_2019(temp = Temp.Cat, q10, a, b, c),
                        data = .x,
                        iter = c(5,5,5,5),
                        start_lower = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'rezende_2019') - 1,
                        start_upper = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'rezende_2019') + 1,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


plot_fit(rezende_fits) + labs(title = "rezende")
get_pars(rezende_fits)

# fit thomas formulation in rTPC
thomas_fits <- Photodata %>% group_by(species) %>% nest() %>%
  mutate(model = map(data, ~nls_multstart(transformed.rate~thomas_2012(temp = Temp.Cat, a, b, c, topt),
                        data = .x,
                        iter = c(5,5,5,5),
                        start_lower = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'thomas_2012') - 1,
                        start_upper = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'thomas_2012') + 1,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


plot_fit(thomas_fits) + labs(title = "thomas")
get_pars(thomas_fits)

# fit thomas2 formulation in rTPC
thomas2_fits <- Photodata %>% group_by(species) %>% nest() %>%
  mutate(model = map(data, ~nls_multstart(transformed.rate~thomas_2017(temp = Temp.Cat, a, b, c, d, e),
                        data = .x,
                        iter = c(5,5,5,5,5),
                        start_lower = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'thomas_2017') - 1,
                        start_upper = get_start_vals(.x$temp, .x$transformed.rate, model_name = 'thomas_2017') + 1,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


plot_fit(thomas2_fits) + labs(title = "thomas2")
get_pars(thomas2_fits)


# Combine model pars
allpars <- bind_rows(.id = "model",
  spain = get_pars(spain_fits),
  hinshelwood = get_pars(hinshelwood_fits),
  joehnk = get_pars(joehnk_fits),
  kamykowski = get_pars(kamykowski_fits),
  lactin2 = get_pars(lactin2_fits),
  lrf = get_pars(lrf_fits),
  rezende = get_pars(rezende_fits),
  thomas = get_pars(thomas_fits),
  thomas2 = get_pars(thomas2_fits)
)


allpars %>%
  pivot_longer(3:12, names_to = "variable", values_to = "value") %>%
  filter(variable %in% c("breadth", "ctmax", "eh", "rmax", "topt")) %>%
  filter(model %in% c("lactin2", "spain", "rezende")) %>%
  ggplot(aes(x = species, y = value, color = model, shape = model)) +
  geom_point() +
  geom_line(aes(group = model, lwd = abs(AIC^3), alpha = 0.3)) +
  facet_wrap(~variable, scales = "free")


allpars %>%
  group_by(model) %>%
  summarise(AIC = sum(AIC)) %>%
  arrange(AIC)

```

```{r fit_rezende_grp}
rezende_fits <- Photodata %>% group_by(species) %>% nest() %>%
  mutate(model = map(data, ~nls_multstart(transformed.rate~rezende_2019(temp = Temp.Cat, q10, a, b, c),
                        data = .x,
                        iter = c(5,5,5,5),
                        start_lower = get_start_vals(.x$temp, .x$transformed.rate, 
                                                     model_name = 'rezende_2019') - 1,
                        start_upper = get_start_vals(.x$temp, .x$transformed.rate, 
                                                     model_name = 'rezende_2019') + 1,
                        supp_errors = 'Y',
                        convergence_count = FALSE)))


plot_fit(rezende_fits) + labs(title = "rezende")
get_pars(rezende_fits)

get_pars(rezende_fits) %>%
  write_csv(path = "data/processed/TPC_Dec_Photo_params_rezende_grp.csv")
```

```{r}
# Bootstrapping

## Refit models using nlsLM and calculate parameters
d_fits_groups2 <- rezende_fits %>%
  mutate(fit_nlsLM = map2(data, model, 
                          ~minpack.lm::nlsLM(transformed.rate~rezende_2019(temp = Temp.Cat, q10, a, b, c),
                                             data = .x,
                                             start = coef(.y),
                                             lower = low_lims,
                                             upper = upper_lims,
                                             weights = rep(1, times = nrow(.x)))),
         params = map(fit_nlsLM, ~calc_params(.)))

# Calculate confidence intervals on parameters by bootstrapping

### P. grandis [[1]]
fit_nlsLM_Pgra <- minpack.lm::nlsLM(transformed.rate~rezende_2019(temp = Temp.Cat, q10, a, b, c),
                                             data = rezende_fits$data[[1]],
                                             start = coef(rezende_fits$model[[1]]),
                                             lower = low_lims,
                                             upper = upper_lims,
                                             weights = rep(1, times = nrow(rezende_fits$data[[1]])))

ci_extra_params_Pgra <- car::Boot(fit_nlsLM_Pgra, f = function(x){unlist(calc_params(x))}, 
                                 labels = names(calc_params(fit_nlsLM_Pgra)), 
                                 R = 200, method = 'case') %>%
  confint(., method = 'bca', level = 0.9) %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')

### P. verrucosa [[1]]
fit_nlsLM_Pver <- minpack.lm::nlsLM(transformed.rate~rezende_2019(temp = Temp.Cat, q10, a, b, c),
                                             data = rezende_fits$data[[2]],
                                             start = coef(rezende_fits$model[[2]]),
                                             lower = low_lims,
                                             upper = upper_lims,
                                             weights = rep(1, times = nrow(rezende_fits$data[[2]])))

ci_extra_params_Pver <- car::Boot(fit_nlsLM_Pver, f = function(x){unlist(calc_params(x))}, 
                             labels = names(calc_params(fit_nlsLM_Pver)), 
                             R = 200, method = 'case') %>%
  confint(., method = 'bca', level = 0.9) %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')


ci_extra_params <- bind_rows(.id = "species",
                             `P. grandis` = ci_extra_params_Pgra, 
                             `P. verrucosa` = ci_extra_params_Pver) 

par_est <- rezende_fits %>%
  mutate(params = map(model, ~calc_params(.))) %>%
  select(species, params) %>%
  unnest(params) %>%
  pivot_longer(2:12, names_to = "param")

res <- left_join(ci_extra_params, par_est)
  
ggplot(res, aes(species, value)) +
  geom_point(size = 1) +
  geom_linerange(aes(ymin = conf_lower, ymax = conf_upper)) +
  theme_bw() +
  facet_wrap(~param, scales = 'free') +
  scale_x_discrete('') +
  labs(title = 'Calculation of confidence intervals for extra parameters',
       subtitle = 'For the bacteria TPC; using case resampling')


write_csv(res, path = "data/processed/TPC_Dec_Photo_params_rezende_grp_confint.csv")
```
